requirements.txt
azure_storage==0.37.0
Flask==2.2.2
openai==0.27.0
PyYAML==6.0
requests==2.28.2
wechatpy==1.8.18
